requirements.txt
langchain
langchain-core
langchain-community

langchain-huggingface
transformers
huggingface-hub

openai
langchain-openai

groq
langchain-groq

python-dotenv

numpy
scikit-learn

pydantic
typing

streamlit

tqdm 
torch
opencv-python

accelerate

numpy